refactor(dm): split PlayerDetailDialog into folder modules
